requirements.txt
async-generator==1.10
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.2
h11==0.13.0
idna==3.3
mypy-extensions==0.4.3
outcome==1.1.0
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-qbittorrent==0.4.2
requests==2.27.1
rich==12.2.0
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
